requirements.txt
<<<<<<< HEAD
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
flatbuffers==23.1.4
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
imageio==2.24.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
six==1.16.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu @ file:///F:/Data/Kuli-ah/Semester%205/Kecerdasan%20Buatan/leaf-classification/tensorflow_cpu-2.11.0-cp39-cp39-win_amd64.whl
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
tifffile==2022.10.10
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
=======
tensorflow-cpu
numpy
pillow
matplotlib
scikit-image
scipy
flask
gunicorn
>>>>>>> 962d838 (update)
